+++ core/partitioning/shape_analysis.cpp
@@ -39,7 +39,7 @@ std::unordered_map<const torch::jit::Value*, torch::jit::IValue> generateRandomI
       std::vector<torch::jit::IValue> list;
       c10::TypePtr elementType = c10::TensorType::get();
       auto generic_list = c10::impl::GenericList(elementType);
-      for (int i = 0; i < input.second.size(); i++) {
+      for (size_t i = 0; i < input.second.size(); i++) {
         auto in = generateSingleInput(input.second[i], types[input.first][i]);
         generic_list.push_back(in.clone());
       }
@@ -47,7 +47,7 @@ std::unordered_map<const torch::jit::Value*, torch::jit::IValue> generateRandomI
     } else if (input.first->type()->kind() == torch::jit::TypeKind::TupleType) {
       // create tuple
       std::vector<torch::jit::IValue> list;
-      for (int i = 0; i < input.second.size(); i++) {
+      for (size_t i = 0; i < input.second.size(); i++) {
         auto in = generateSingleInput(input.second[i], types[input.first][i]);
         list.push_back(in.clone());
       }
@@ -56,7 +56,7 @@ std::unordered_map<const torch::jit::Value*, torch::jit::IValue> generateRandomI
     } else {
       auto in = generateSingleInput(input.second[0], types[input.first][0]);
       ivalue_map[input.first] = in.clone();
-      
+
     }
   }
   return ivalue_map;
